+++ src/main/java/fr/istic/taa/jaxrs/dao/ChildDAO.java
@@ -2,7 +2,9 @@ package fr.istic.taa.jaxrs.dao;
 
 import fr.istic.taa.jaxrs.dao.generic.AbstractJpaDao;
 import fr.istic.taa.jaxrs.domain.Child;
+import fr.istic.taa.jaxrs.exceptions.ValueAlreadyExistsException;
 
+import javax.ws.rs.NotFoundException;
 import java.util.List;
 
 public class ChildDAO extends AbstractJpaDao<Long, Child> {
@@ -11,7 +13,10 @@ public class ChildDAO extends AbstractJpaDao<Long, Child> {
         super(Child.class);
     }
 
-    public void createChild(Child child) {
+    public void createChild(Child child) throws ValueAlreadyExistsException {
+        if (manager.contains(child)) {
+            throw new ValueAlreadyExistsException("The child object is already inserted.");
+        }
         var tx = manager.getTransaction();
         tx.begin();
         manager.persist(child);
@@ -35,17 +40,16 @@ public class ChildDAO extends AbstractJpaDao<Long, Child> {
     }
 
     public void removeChildByName(String firstName, String lastName) {
-        this.manager.createQuery("delete from Child c where c.firstName = :firstName and c.lastName = :lastName")
-                .setParameter("lastName", lastName)
-                .setParameter("firstName", firstName);
-    }
-
-    public void removeAppointmentToPatient(String firstName, String lastName, Long appointmentId) {
-        this.manager.createQuery("delete from Appointment app " +
-                        "where app.id = :id and app.patient.lastName = :lastName " +
-                        "and app.patient.firstName = :firstName")
-                .setParameter("id", appointmentId)
-                .setParameter("lastName", lastName)
-                .setParameter("firstName", firstName);
+        var res = manager.createQuery("select c from Child c where c.firstName like :first and c.lastName like :last", Child.class)
+                .setParameter("first", firstName)
+                .setParameter("last", lastName)
+                .getResultList();
+        if (res.isEmpty()) {
+            throw new NotFoundException("The requested child could not be find.");
+        } else {
+            this.manager.createQuery("delete from Child c where c.firstName = :firstName and c.lastName = :lastName")
+                    .setParameter("lastName", lastName)
+                    .setParameter("firstName", firstName);
+        }
     }
 }
